requirements.txt
Django==1.9.7
pycrypto
eth_account
mysqlclient
requests


